requirements.txt
pandas
torch
marker-pdf
pypandoc
chemdataextractor2